libftfinished

+++ 42cursus/libft/ft_strlcat.c
@@ -6,13 +6,13 @@
 /*   By: aokur <aokur@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/20 14:00:11 by aokur             #+#    #+#             */
-/*   Updated: 2025/06/28 17:09:57 by aokur            ###   ########.fr       */
+/*   Updated: 2025/07/09 16:01:04 by aokur            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned	int	ft_strlcat(char *dest, char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, char *src, size_t size)
 {
 	unsigned int	i;
 	unsigned int	dest_len;
@@ -24,7 +24,7 @@ unsigned	int	ft_strlcat(char *dest, char *src, unsigned int size)
 	if (size <= dest_len)
 		return (src_len + size);
 	while (src[i] != '\0' && (dest_len + i < size - 1))
-	{	
+	{
 		dest[dest_len + i] = src[i];
 		i++;
 	}
