+++ 42cursus/libft/libft.h
@@ -6,18 +6,15 @@
 /*   By: aokur <aokur@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/29 19:06:56 by aokur             #+#    #+#             */
-/*   Updated: 2025/07/07 16:12:28 by aokur            ###   ########.fr       */
+/*   Updated: 2025/07/09 16:17:49 by aokur            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
-# include <stdio.h>
-# include <stddef.h>
 # include <unistd.h>
 # include <stdlib.h>
-# include <fcntl.h>
 
 int			ft_atoi(const char *str);
 void		ft_bzero(void *s, size_t n);
@@ -33,8 +30,8 @@ void		*ft_memcpy(void *dest, const void *src, size_t n);
 void		*ft_memmove(void *dst, const void *src, size_t len);
 void		*ft_memset(void *b, int c, size_t len);
 char		*ft_strchr(const char *s, int c);
-unsigned	int	ft_strlcat(char *dest, char *src, unsigned int size);
-unsigned	int	ft_strlcpy(char *dest, char *src, unsigned int size);
+size_t		ft_strlcat(char *dest, char *src, size_t size);
+size_t		ft_strlcpy(char *dest, char *src, size_t size);
 size_t		ft_strlen(const char *s);
 int			ft_strncmp(const char *s1, const char *s2, size_t n);
 char		*ft_strnstr(const char *haystack, const char *needle, size_t len);
